moved createKernelArguments into ConfiguredKernel; moved cuda library interfaces into separate package
