+++ projects/com.nvidia.grcuda/src/com/nvidia/grcuda/gpu/Kernel.java
@@ -130,151 +130,8 @@ public class Kernel implements TruffleObject {
         return kernelComputationArguments;
     }
 
-    KernelArguments createKernelArguments(Object[] args, InteropLibrary booleanAccess,
-                    InteropLibrary int8Access, InteropLibrary int16Access,
-                    InteropLibrary int32Access, InteropLibrary int64Access, InteropLibrary doubleAccess)
-                    throws UnsupportedTypeException, ArityException {
-        if (args.length != kernelComputationArguments.length) {
-            CompilerDirectives.transferToInterpreter();
-            throw ArityException.create(kernelComputationArguments.length, args.length);
-        }
-        KernelArguments kernelArgs = new KernelArguments(args, this.kernelComputationArguments);
-        for (int paramIdx = 0; paramIdx < kernelComputationArguments.length; paramIdx++) {
-            Object arg = args[paramIdx];
-            ComputationArgument param = kernelComputationArguments[paramIdx];
-            Type paramType = param.getType();
-            try {
-                if (param.isPointer()) {
-                    if (arg instanceof DeviceArray) {
-                        DeviceArray deviceArray = (DeviceArray) arg;
-                        if (!param.isSynonymousWithPointerTo(deviceArray.getElementType())) {
-                            throw new GrCUDAException("device array of " + deviceArray.getElementType() + " cannot be used as pointer argument " + paramType);
-                        }
-                        UnsafeHelper.PointerObject pointer = UnsafeHelper.createPointerObject();
-                        pointer.setValueOfPointer(deviceArray.getPointer());
-                        kernelArgs.setArgument(paramIdx, pointer);
-                    } else if (arg instanceof MultiDimDeviceArray) {
-                        MultiDimDeviceArray deviceArray = (MultiDimDeviceArray) arg;
-                        if (!param.isSynonymousWithPointerTo(deviceArray.getElementType())) {
-                            throw new GrCUDAException("multi-dimensional device array of " +
-                                            deviceArray.getElementType() + " cannot be used as pointer argument " + paramType);
-                        }
-                        UnsafeHelper.PointerObject pointer = UnsafeHelper.createPointerObject();
-                        pointer.setValueOfPointer(deviceArray.getPointer());
-                        kernelArgs.setArgument(paramIdx, pointer);
-                    } else {
-                        CompilerDirectives.transferToInterpreter();
-                        throw UnsupportedTypeException.create(new Object[]{arg}, "expected DeviceArray type");
-                    }
-                } else {
-                    // by-value argument
-                    switch (paramType) {
-                        case BOOLEAN: {
-                            UnsafeHelper.Integer8Object int8 = UnsafeHelper.createInteger8Object();
-                            int8.setValue(booleanAccess.asBoolean(arg) ? ((byte) 1) : ((byte) 0));
-                            kernelArgs.setArgument(paramIdx, int8);
-                            break;
-                        }
-                        case SINT8:
-                        case CHAR: {
-                            UnsafeHelper.Integer8Object int8 = UnsafeHelper.createInteger8Object();
-                            int8.setValue(int8Access.asByte(arg));
-                            kernelArgs.setArgument(paramIdx, int8);
-                            break;
-                        }
-                        case SINT16: {
-                            UnsafeHelper.Integer16Object int16 = UnsafeHelper.createInteger16Object();
-                            int16.setValue(int16Access.asShort(arg));
-                            kernelArgs.setArgument(paramIdx, int16);
-                            break;
-                        }
-                        case SINT32:
-                        case WCHAR: {
-                            UnsafeHelper.Integer32Object int32 = UnsafeHelper.createInteger32Object();
-                            int32.setValue(int32Access.asInt(arg));
-                            kernelArgs.setArgument(paramIdx, int32);
-                            break;
-                        }
-                        case SINT64:
-                        case SLL64:
-                            // no larger primitive type than long -> interpret long as unsigned
-                        case UINT64:
-                        case ULL64: {
-                            UnsafeHelper.Integer64Object int64 = UnsafeHelper.createInteger64Object();
-                            int64.setValue(int64Access.asLong(arg));
-                            kernelArgs.setArgument(paramIdx, int64);
-                            break;
-                        }
-                        case UINT8:
-                        case CHAR8: {
-                            int uint8 = int16Access.asShort(arg);
-                            if (uint8 < 0 || uint8 > 0xff) {
-                                CompilerDirectives.transferToInterpreter();
-                                throw createExceptionValueOutOfRange(paramType, uint8);
-                            }
-                            UnsafeHelper.Integer8Object int8 = UnsafeHelper.createInteger8Object();
-                            int8.setValue((byte) (0xff & uint8));
-                            kernelArgs.setArgument(paramIdx, int8);
-                            break;
-                        }
-                        case UINT16:
-                        case CHAR16: {
-                            int uint16 = int32Access.asInt(arg);
-                            if (uint16 < 0 || uint16 > 0xffff) {
-                                CompilerDirectives.transferToInterpreter();
-                                throw createExceptionValueOutOfRange(paramType, uint16);
-                            }
-                            UnsafeHelper.Integer16Object int16 = UnsafeHelper.createInteger16Object();
-                            int16.setValue((short) (0xffff & uint16));
-                            kernelArgs.setArgument(paramIdx, int16);
-                            break;
-                        }
-                        case UINT32: {
-                            long uint32 = int64Access.asLong(arg);
-                            if (uint32 < 0 || uint32 > 0xffffffffL) {
-                                CompilerDirectives.transferToInterpreter();
-                                throw createExceptionValueOutOfRange(paramType, uint32);
-                            }
-                            UnsafeHelper.Integer32Object int32 = UnsafeHelper.createInteger32Object();
-                            int32 = UnsafeHelper.createInteger32Object();
-                            int32.setValue((int) (0xffffffffL & uint32));
-                            kernelArgs.setArgument(paramIdx, int32);
-                            break;
-                        }
-                        case FLOAT: {
-                            UnsafeHelper.Float32Object fp32 = UnsafeHelper.createFloat32Object();
-                            // going via "double" to allow floats to be initialized with doubles
-                            fp32.setValue((float) doubleAccess.asDouble(arg));
-                            kernelArgs.setArgument(paramIdx, fp32);
-                            break;
-                        }
-                        case DOUBLE: {
-                            UnsafeHelper.Float64Object fp64 = UnsafeHelper.createFloat64Object();
-                            fp64.setValue(doubleAccess.asDouble(arg));
-                            kernelArgs.setArgument(paramIdx, fp64);
-                            break;
-                        }
-                        default:
-                            CompilerDirectives.transferToInterpreter();
-                            throw UnsupportedTypeException.create(new Object[]{arg},
-                                            "unsupported by-value parameter type: " + paramType);
-                    }
-                }
-            } catch (UnsupportedMessageException e) {
-                CompilerDirectives.transferToInterpreter();
-                throw UnsupportedTypeException.create(new Object[]{arg},
-                                "expected type " + paramType + " in argument " + arg);
-            }
-        }
-        return kernelArgs;
-    }
-
-    private static GrCUDAException createExceptionValueOutOfRange(Type type, long value) {
-        return new GrCUDAException("value " + value + " is out of range for type " + type);
-    }
-
-    public long getKernelFunctionHandle(int deviceId) {
-        if (modules.get(deviceId).isClosed()) {
+    public long getKernelFunctionHandle() {
+        if (module.isClosed()) {
             CompilerDirectives.transferToInterpreter();
             throw new GrCUDAException("CUmodule containing kernel " + kernelName + " is already closed");
         }
